requirements.txt
python-telegram-bot>=20.0
langchain>=0.0.90
ollama>=0.1.0